Implement routing with redux-simple-router
This is more complex change because now the individual page comments
connect to the Redux store, instead of this being handled in App. The
main reducer now only looks at the messages key of the overall state,
since we're also storing routing information.

I also reworked the server-side rendering to include the request path.